Numeric binary operators implemented
The bytecode generator is going to be a large file. It could do with more modularity. I could simply split it into node handling classes like the VM does, but we have the disadvantage that we need all of these function definitions in order to get static double dispatch. The compiler doesn't really need to be fast, though, it may be a maintainability boon to lose those guarantees. I'll have to think about it.