Implemented foreach style loop
The loop body is a function, which may end up being a bad move. We'll see. Functions have access to all non-shadowed bindings in their enclosing scopes in technical terms, and it's going to be up to the typechecker to constrain the scope further, which it won't do here because it has no idea that loop bodies are functions (because they aren't, to it).

It did force my hand on implementing function generation too. I'm still not convinced by this design. Emitting the opcodes for function definitions is deferred until the last step, so we can resolve things which need global knowledge like function position references. Whenever we find a function we just add it to the function table, which creates a const table entry for the function pointer. When finalising, each function is actually compiled and we learn where we're going to put it within the instruction stream, at which point the const table entry can be updated. "Const table" is actually a bit of a misnomer for the code generator, though it is correct for the runtime.