TypeClass kinda supported
Typeclasses now /compile/ the correct functions, and we have the information available to know which overload is which. What we don't, however, have, is accurate type information at the call site. f(x) can be dispatched because we probably know what x is, or could look it up. f(5) we can't really do, f(g(x)) is becoming a lot more complex. The code generator discards type information as it goes, so we can't really make these choices here. As such, while we generate the best code we can, we can't generate the correct function calls until the type checker is informing our overload resolution, which it can't currently do because it doesn't actually exist yet.