Compound Constraint rules mostly implemented
Major missing feature is monobound ranges aren't currently supertypes of polybound ranges when they should be. This is because it's kinda awkward to implement and I haven't thought of an elegant way yet. The brute force way would just be to write in support on a case by case basis, but that doesn't scale well. The entire architecture may want to be reworked.