And and Or Composite constraints implemented
These are interesting constraints, because what they capture is surprisingly complex. Or is simpler, but even then. The supertype relation in or is true if any option is a supertype, *or* if any option is equal or assignable. Whereas normally this wouldn't be the case, in an Or it is because one equal type plus another type still leaves the Or type with a larger set of values, making it a super type, even though the other values in that set are potentially entirely disjoint!