Function call information exporting
I initially tried to annotate the tree, against my better judgement, but found it too distasteful and too stateful. The type checker now just exports a set of function definitions and their assigned types, and a set of function calls and their assigned types. This should be enough for the code generator to match the two up.