Half-implemented record types; implemented complex datatype checking
Complex datatypes currently have assignability and subtyping rules which make a complex object where every attribute is assignable to another complex object's attribute assignable itself. This may be undesirable, I'm not sure. It may just not get used by the typechecker itself.

Complex objects, regardless, are represented by a collection of functions and a type. These functions are compiler generated so I can't assign them to a definition node, so overloading could be difficult. I may just not allow it, which would cause a slight pain naming but also would reduce the astonishment factor of having one function which is both a regular function and either a ctor or accessor!