Implemented record typechecking
Records in the typechecker work much like records in the code generator. They make a bunch of magic functions. In this case the functions themselves don't even exist, there's only the minimum required to typecheck at the boundaries. This does mean that the types don't "really" exist, but I don't think that should be a problem, as no types "really" exist to the code generator. Function overload resolution will require dealing with type references that don't go anywhere, but I don't think the code generator will need to follow them anyway. All it's doing is matching things up, it could do that on a number if it wanted.