Array typechecking improved, and a problem
I can't do typeclasses. Not sensibly, anyway. A function which accepts a typeclass needs to be able to call functions which differ depending on what the parameters are. I have no dynamic function call ability, so I'd need to compile a seperate function for every usage. This is doable, but silly. Better would be to implement dynamic function calls and transform functions which take a typeclass to also accept a method table per class, which is absolutely a thing that could be done, but would require some fairly significant re-engineering that I don't want to do right now. I'm going to drop typeclasses for now and come back to them later, when I know I have the time to mess around with architecture.