Basic parser combinator framework implemented
That was interesting. There were a couple of neat decisions to be made there. Firstly, the parser returning a list of matches is there to allow parsed entities to "build up". A traditional combinator here would build up a list of tokens that that parser recognises, which could later be transformed into a more structured object. I decided to take a different approach and build a list of partial trees, where the types are variable. A parser that makes AST nodes and a parser that recognises tokens can live together perfectly fine so long as the output of one can be transformed into the other--the tokens can always either become nodes, or be ignored, so this works out. The upshot of this is that the parser is going to have three stages, not two, with the last two stages existing together. The first stage is a lexing parse, and the second stage does syntactic resolution and AST transformation as mostly separate, but entwined, stages.