All parser tests pass
So I just rewrote this as a little LLRD. It did not take very long.

There were two notable changes to the syntax. Firstly, my program type is now a block type, which only allows statements. This means that you cannot, at a syntax level, have function defs or record defs or whatever anywhere but top-level. I need a top-level program type, which I will do in a moment.

Secondly, record types now use brackets instead of braces. This is a teeny choice that slightly simplifies the grammar and parser with no loss of comprehensibility, so I'm going to call it a tiny net gain.