Now outputs runnable bytecode
A few changes were necessary to bring this into line with reality. The bytecode for for had a few issues, the stack would grow infinitely because function calls should clean up their own stack, and I forgot this. So function calls now emit pop_tops for their parameters, too.

We also have our first library function, putc! This writes a character to the terminal. A better method of library functions needs to exist, but this is a basic proof of concept. It's even typechecked!